codie/analysis_orchestrator/requirements.txt
fastapi
uvicorn
pydantic
httpx
websockets
gitpython
requests